+++ Repositories/Presenters/WishPresenter.cs
@@ -38,38 +38,87 @@ namespace Repositories
         }
         public void ShowTable(bool sort = false)
         {
-            var clients = GetClients();
-
-            dgvElements = wishRepository.GetTable();
-            dgv.Rows.Clear();
-            foreach (DBWish wish in dgvElements)
+            try
             {
-                //int id, int client, string township, string apartamentOrHouse, int area, int numberOfRooms, int cost
-                DBPerson client = null;
-                clients.TryGetValue(wish.client, out client);
+                var clients = GetClients();
 
-                dgv.Rows.Add(wish.id,
-                    String.Format("[{0}] {1} {2}", client.id, client.surname, client.name),
-                    wish.township,
-                    wish.apartamentOrHouse,
-                    wish.area.ToString(),
-                    wish.numberOfRooms.ToString(),
-                    wish.cost);
+                dgvElements = wishRepository.GetTable();
+                dgv.Rows.Clear();
+                foreach (DBWish wish in dgvElements)
+                {
+                    //int id, int client, string township, string apartamentOrHouse, int area, int numberOfRooms, int cost
+                    DBPerson client = null;
+                    clients.TryGetValue(wish.client, out client);
+
+                    dgv.Rows.Add(wish.id,
+                        String.Format("[{0}] {1} {2}", client.id, client.surname, client.name),
+                        wish.township,
+                        wish.apartamentOrHouse,
+                        wish.area.ToString(),
+                        wish.numberOfRooms.ToString(),
+                        wish.cost);
+                }
             }
+            catch (Exception) { MessageBox.Show("Ошибка базы данных.", "Ошибка", MessageBoxButtons.OK, MessageBoxIcon.Error); }
+
             if (sort)
                 dgv.Sort(dgv.Columns[0], ListSortDirection.Ascending);
         }
-        public void AddToTable(DBWish wish)
+        public bool AddToTable(DBWish wish)
         {
-            wishRepository.AddToTable(wish);
+            List<string> errorList;
+            bool checkFlag = WishValidator.checkAddition(wish, out errorList);
+            try
+            {
+                if (checkFlag)
+                    wishRepository.AddToTable(wish);
+            }
+            catch (Exception) { errorList.Add("Ошибка базы данных."); }
+
+            ShowErrors(errorList);
+
+            return checkFlag;
         }
-        public void UpdateTable(DBWish wish)
+        public bool UpdateTable(DBWish wish)
         {
-            wishRepository.UpdateTable(wish);
+            List<string> errorList;
+            bool checkFlag = WishValidator.checkAddition(wish, out errorList);
+            try
+            {
+                if (checkFlag)
+                    wishRepository.UpdateTable(wish);
+            }
+            catch (Exception) { errorList.Add("Ошибка базы данных."); }
+
+            ShowErrors(errorList);
+
+            return checkFlag;
         }
-        public void DeleteFromTable(int id)
+        public bool DeleteFromTable(int id)
         {
-            wishRepository.DeleteFromTable(id);
+            List<string> errorList;
+            bool checkFlag = WishValidator.checkDelete(id, out errorList);
+            try
+            {
+                if (checkFlag)
+                    wishRepository.DeleteFromTable(id);
+            }
+            catch (Exception) { errorList.Add("Ошибка базы данных."); }
+
+            ShowErrors(errorList);
+
+            return checkFlag;
+        }
+
+        private void ShowErrors(List<string> errorList)
+        {
+            if (errorList.Count == 0)
+                return;
+
+            string errors = "";
+            foreach (string s in errorList)
+                errors += s + System.Environment.NewLine;
+            MessageBox.Show(errors, "Ошибка", MessageBoxButtons.OK, MessageBoxIcon.Error);
         }
     }
 }
